+++ src/components/CardItem/CardItem.tsx
@@ -1,24 +1,22 @@
 import React from 'react';
 import './CardItem.css';
 import { Col, Card } from 'antd';
-import { PokemonDto } from '../../interfaces/PokemonDto.tsx';
+import { CardShortResponse } from '../../models/ResponseModels/CardShortResponse.tsx';
 
 const { Meta } = Card;
 
-export const CardItem: React.FC<{ pokemon: PokemonDto }> = ({ pokemon }) => (
+export const CardItem: React.FC<{ card: CardShortResponse }> = ({ card }) => (
   <Col span={3}>
     <Card
       hoverable={true}
-      onClick={() => window.location.assign('/pokemon/' + pokemon.id)}
+      onClick={() => window.location.assign('/card/' + card.id)}
       className="pokemon-card"
-      title={pokemon.name}
+      title={card.name}
       bordered={false}
-      cover={
-        <img className="img" alt={pokemon.id} src={pokemon.images.small} />
-      }
+      cover={<img className="img" alt={card.id} src={card.images!.small} />}
     >
-      {pokemon.rarity}
-      <Meta title={pokemon.artist} description={pokemon.flavorText} />
+      {card.rarity}
+      <Meta title={card.artist} description={card.flavorText} />
     </Card>
   </Col>
 );
