--- src/interfaces/PokemonDto.tsx
@@ -1,20 +1,0 @@
-export interface PokemonDto {
-  id: string;
-  name: string;
-  supertype: string;
-  subtypes: string[];
-  hp: string;
-  types: string[];
-  evolvesTo: string[];
-  rules: string[];
-  convertedRetreatCost: number;
-  number: string;
-  artist: string;
-  rarity: string;
-  flavorText: string;
-  nationalPokedexNumbers: number[];
-  images: {
-    large: string;
-    small: string;
-  };
-}
