+++ src/pages/PokemonCardPage/PokemonCardPage.scss
@@ -1,0 +1,22 @@
+.card-page {
+  margin: 20px 0;
+  display: flex;
+  justify-content: center;
+
+  .card-page-container {
+    display: flex;
+    gap: 10px;
+    flex-direction: column;
+    align-items: center;
+
+    .ant-card-cover {
+      display: flex;
+      justify-content: center;
+
+      img {
+        width: auto;
+        height: 50vh;
+      }
+    }
+  }
+}
